AIFiles/requirements (3).txt
fastapi
uvicorn
diffusers
torch
transformers
pillow
scipy
